# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for مولدتي - Iraqi Generator Management System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  owner: "owner",
  admin: "admin",
  worker: "worker",
} as const;

export interface User {
  id: number;
  name: string;
  username: string;
  role: UserRole;
  assignedGeneratorId?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token?: string;
}

export interface CreateUserBody {
  name: string;
  username: string;
  password: string;
  role: UserRole;
  assignedGeneratorId?: number | null;
}

export interface UpdateUserBody {
  name?: string;
  username?: string;
  password?: string;
  role?: UserRole;
  assignedGeneratorId?: number | null;
}

export interface Generator {
  id: number;
  name: string;
  location: string;
  /** Price per ampere in Iraqi Dinars */
  pricePerAmpereIqd: number;
  ownerId: number;
  activeSubscribersCount?: number;
  createdAt: string;
}

export interface CreateGeneratorBody {
  name: string;
  location: string;
  pricePerAmpereIqd: number;
}

export interface UpdateGeneratorBody {
  name?: string;
  location?: string;
  pricePerAmpereIqd?: number;
}

export type SubscriberStatus =
  (typeof SubscriberStatus)[keyof typeof SubscriberStatus];

export const SubscriberStatus = {
  active: "active",
  paused: "paused",
} as const;

export interface Subscriber {
  id: number;
  name: string;
  breakerOwnerName: string;
  phoneNumber?: string | null;
  generatorId: number;
  defaultAmperes: number;
  status: SubscriberStatus;
  /** Current outstanding debt in IQD */
  currentDebt: number;
  createdAt: string;
}

export interface CreateSubscriberBody {
  name: string;
  breakerOwnerName: string;
  phoneNumber?: string | null;
  generatorId: number;
  defaultAmperes: number;
  status?: SubscriberStatus;
}

export interface UpdateSubscriberBody {
  name?: string;
  breakerOwnerName?: string;
  phoneNumber?: string | null;
  defaultAmperes?: number;
  status?: SubscriberStatus;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export interface Invoice {
  id: number;
  subscriberId: number;
  generatorId: number;
  workerId?: number | null;
  amperes: number;
  previousDebt: number;
  /** (Amperes * Price) + Previous Debt in IQD */
  amountExpected: number;
  amountReceived: number;
  /** YYYY-MM format */
  monthYear: string;
  status: InvoiceStatus;
  notes?: string | null;
  createdAt: string;
  subscriber?: Subscriber | null;
  worker?: User | null;
}

export interface CreateInvoiceBody {
  subscriberId: number;
  generatorId: number;
  amperes: number;
  previousDebt?: number;
  amountReceived: number;
  monthYear: string;
  notes?: string | null;
}

export interface UpdateInvoiceBody {
  amperes?: number;
  amountReceived?: number;
  notes?: string | null;
  status?: InvoiceStatus;
}

export interface BulkGenerateBody {
  generatorId: number;
  /** YYYY-MM format */
  monthYear: string;
}

export interface BulkGenerateResult {
  generated: number;
  skipped: number;
  monthYear: string;
  invoices: Invoice[];
}

export interface Expense {
  id: number;
  generatorId: number;
  expenseType: string;
  /** Amount in IQD */
  amount: number;
  date: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateExpenseBody {
  generatorId: number;
  expenseType: string;
  amount: number;
  date: string;
  notes?: string | null;
}

export interface UpdateExpenseBody {
  expenseType?: string;
  amount?: number;
  date?: string;
  notes?: string | null;
}

export interface DashboardSummary {
  totalExpectedRevenue: number;
  totalCollected: number;
  totalOutstanding: number;
  totalExpenses: number;
  netProfit: number;
  activeSubscribersCount: number;
  paidCount: number;
  partialCount: number;
  unpaidCount: number;
  collectionRate: number;
}

export interface MonthlyFinancial {
  monthYear: string;
  revenue: number;
  expenses: number;
  netProfit: number;
  invoiceCount: number;
  paidCount: number;
}

export interface TransactionRecord {
  id: number;
  subscriberName: string;
  breakerOwnerName: string;
  generatorName: string;
  workerName?: string | null;
  amountExpected: number;
  amountReceived: number;
  status: InvoiceStatus;
  monthYear: string;
  createdAt: string;
}

export interface InvoiceStatusBreakdown {
  paid: number;
  partial: number;
  unpaid: number;
  totalAmount: number;
  collectedAmount: number;
}

export type ListSubscribersParams = {
  generatorId?: number;
  /**
   * Search by name or breaker owner name
   */
  search?: string;
  status?: SubscriberStatus;
};

export type ListInvoicesParams = {
  generatorId?: number;
  subscriberId?: number;
  workerId?: number;
  /**
   * Month filter e.g. 2025-01
   */
  monthYear?: string;
  status?: InvoiceStatus;
};

export type ListExpensesParams = {
  generatorId?: number;
  /**
   * Month filter e.g. 2025-01
   */
  month?: string;
};

export type GetDashboardSummaryParams = {
  generatorId?: number;
  monthYear?: string;
};

export type GetFinancialHistoryParams = {
  generatorId?: number;
  year?: number;
};

export type GetRecentTransactionsParams = {
  generatorId?: number;
  limit?: number;
};

export type GetInvoiceStatusBreakdownParams = {
  generatorId?: number;
  monthYear?: string;
};