# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مولدتي - Iraqi Generator Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /generators:
    get:
      operationId: listGenerators
      tags: [generators]
      summary: List generators
      responses:
        "200":
          description: List of generators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Generator"
    post:
      operationId: createGenerator
      tags: [generators]
      summary: Create a new generator (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeneratorBody"
      responses:
        "201":
          description: Generator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generator"
  /generators/{id}:
    get:
      operationId: getGenerator
      tags: [generators]
      summary: Get generator by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generator found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generator"
    put:
      operationId: updateGenerator
      tags: [generators]
      summary: Update generator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGeneratorBody"
      responses:
        "200":
          description: Generator updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generator"
    delete:
      operationId: deleteGenerator
      tags: [generators]
      summary: Delete generator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Generator deleted
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses with filters
      parameters:
        - name: generatorId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
          description: "Month filter e.g. 2025-01"
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Expense deleted
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary - totals, active subscribers, unpaid count
      parameters:
        - name: generatorId
          in: query
          schema:
            type: integer
        - name: monthYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/financial-history:
    get:
      operationId: getFinancialHistory
      tags: [dashboard]
      summary: Get monthly financial history - revenue, expenses, net profit
      parameters:
        - name: generatorId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly financial history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyFinancial"
  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get recent invoice transactions with worker info
      parameters:
        - name: generatorId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionRecord"
  /dashboard/invoice-status-breakdown:
    get:
      operationId: getInvoiceStatusBreakdown
      tags: [dashboard]
      summary: Get breakdown of invoices by status for a month
      parameters:
        - name: generatorId
          in: query
          schema:
            type: integer
        - name: monthYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Status breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceStatusBreakdown"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
    UserRole:
      type: string
      enum: [owner, admin, worker]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        assignedGeneratorId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - username
        - role
        - createdAt
...
    Generator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        pricePerAmpereIqd:
          type: integer
          description: Price per ampere in Iraqi Dinars
        ownerId:
          type: integer
        activeSubscribersCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - pricePerAmpereIqd
        - ownerId
        - createdAt
    CreateGeneratorBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        pricePerAmpereIqd:
          type: integer
      required:
        - name
        - location
        - pricePerAmpereIqd
    UpdateGeneratorBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        pricePerAmpereIqd:
          type: integer
...
    InvoiceStatus:
      type: string
      enum: [paid, partial, unpaid]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        generatorId:
          type: integer
        expenseType:
          type: string
        amount:
          type: integer
          description: Amount in IQD
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - generatorId
        - expenseType
        - amount
        - date
        - createdAt
    CreateExpenseBody:
      type: object
      properties:
        generatorId:
          type: integer
        expenseType:
          type: string
        amount:
          type: integer
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      required:
        - generatorId
        - expenseType
        - amount
        - date
    UpdateExpenseBody:
      type: object
      properties:
        expenseType:
          type: string
        amount:
          type: integer
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
    DashboardSummary:
      type: object
      properties:
        totalExpectedRevenue:
          type: integer
        totalCollected:
          type: integer
        totalOutstanding:
          type: integer
        totalExpenses:
          type: integer
        netProfit:
          type: integer
        activeSubscribersCount:
          type: integer
        paidCount:
          type: integer
        partialCount:
          type: integer
        unpaidCount:
          type: integer
        collectionRate:
          type: number
          format: double
      required:
        - totalExpectedRevenue
        - totalCollected
        - totalOutstanding
        - totalExpenses
        - netProfit
        - activeSubscribersCount
        - paidCount
        - partialCount
        - unpaidCount
        - collectionRate
    MonthlyFinancial:
      type: object
      properties:
        monthYear:
          type: string
        revenue:
          type: integer
        expenses:
          type: integer
        netProfit:
          type: integer
        invoiceCount:
          type: integer
        paidCount:
          type: integer
      required:
        - monthYear
        - revenue
        - expenses
        - netProfit
        - invoiceCount
        - paidCount
    TransactionRecord:
      type: object
      properties:
        id:
          type: integer
        subscriberName:
          type: string
        breakerOwnerName:
          type: string
        generatorName:
          type: string
        workerName:
          type: string
          nullable: true
        amountExpected:
          type: integer
        amountReceived:
          type: integer
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        monthYear:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberName
        - breakerOwnerName
        - generatorName
        - amountExpected
        - amountReceived
        - status
        - monthYear
        - createdAt
    InvoiceStatusBreakdown:
      type: object
      properties:
        paid:
          type: integer
        partial:
          type: integer
        unpaid:
          type: integer
        totalAmount:
          type: integer
        collectedAmount:
          type: integer
      required:
        - paid
        - partial
        - unpaid
        - totalAmount
        - collectedAmount